+++ src/main/java/twilightforest/entity/boss/EntityTFNaga.java
@@ -204,6 +204,7 @@ public class EntityTFNaga extends EntityMob implements IEntityMultiPart {
 		public void startExecuting() {
 			// NAGA SMASH!
 			if (!taskOwner.getWorld().isRemote) {
+
 				AxisAlignedBB bb = taskOwner.getEntityBoundingBox();
 				int minx = MathHelper.floor(bb.minX - 0.75D);
 				int miny = MathHelper.floor(bb.minY + 1.01D);
@@ -211,13 +212,20 @@ public class EntityTFNaga extends EntityMob implements IEntityMultiPart {
 				int maxx = MathHelper.floor(bb.maxX + 0.75D);
 				int maxy = MathHelper.floor(bb.maxY + 0.0D);
 				int maxz = MathHelper.floor(bb.maxZ + 0.75D);
+
 				if (taskOwner.getWorld().isAreaLoaded(new BlockPos(minx, miny, minz), new BlockPos(maxx, maxy, maxz))) {
 					for (int dx = minx; dx <= maxx; dx++) {
 						for (int dy = miny; dy <= maxy; dy++) {
 							for (int dz = minz; dz <= maxz; dz++) {
+
 								BlockPos pos = new BlockPos(dx, dy, dz);
-								if (taskOwner.getWorld().getBlockState(pos).getBlockHardness(taskOwner.getWorld(), pos) >= 0)
+								IBlockState state = taskOwner.getWorld().getBlockState(pos);
+
+								if (state.getBlockHardness(taskOwner.getWorld(), pos) >= 0
+										&& state.getBlock().canEntityDestroy(state, taskOwner.getWorld(), pos, taskOwner)) {
+
 									taskOwner.getWorld().destroyBlock(pos, true);
+								}
 							}
 						}
 					}
@@ -401,8 +409,9 @@ public class EntityTFNaga extends EntityMob implements IEntityMultiPart {
 						for (int dz = minz; dz <= maxz; dz++) {
 							BlockPos pos = new BlockPos(dx, dy, dz);
 							IBlockState state = this.getWorld().getBlockState(pos);
-							if (state.getMaterial() == Material.LEAVES && state.getBlockHardness(this.getWorld(), pos) >= 0)
+							if (state.getMaterial() == Material.LEAVES && state.getBlockHardness(this.getWorld(), pos) >= 0 && state.getBlock().canEntityDestroy(state, this.getWorld(), pos, this)) {
 								this.getWorld().destroyBlock(pos, true);
+							}
 						}
 					}
 				}
@@ -578,7 +587,7 @@ public class EntityTFNaga extends EntityMob implements IEntityMultiPart {
 			BlockPos pos = new BlockPos(dx, dy, dz);
 			IBlockState state = world.getBlockState(pos);
 
-			if (state.getBlockHardness(world, pos) >= 0.0F && !state.getBlock().isAir(state, world, pos)) {
+			if (state.getBlockHardness(world, pos) >= 0.0F && !state.getBlock().isAir(state, world, pos) && state.getBlock().canEntityDestroy(state, world, pos, this)) {
 				// todo limit what can be broken
 				world.destroyBlock(pos, true);
